requirements.txt
# Gmail to Discord Relay Bot - Dependencies

# Gmail API
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0

# Discord API
discord.py==2.3.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1